+++ viewmodel-core/src/commonMain/kotlin/viewmodel/ViewModelUtils.kt
@@ -10,9 +10,9 @@ internal val <T : Any> T.toDetailedString: String
         else -> toString()
     }
 
-fun <I> VModel<I, *>.log(intent: I) = log("Sending Intent ${intent?.toDetailedString}")
+fun <I> ViewModel<I, *>.log(intent: I) = log("Sending Intent ${intent?.toDetailedString}")
 
-fun VModel<*, *>.log(msg: String) = when {
+fun ViewModel<*, *>.log(msg: String) = when {
     msg.contains("error", ignoreCase = true) -> logger.error(msg)
     msg.contains("fail", ignoreCase = true) -> logger.failure(msg)
     else -> logger.info(msg)