+++ viewmodel-test/viewmodel-test-core/src/commonMain/kotlin/viewmodel/viewmodel.test.kt
@@ -3,9 +3,9 @@ package viewmodel
 import kotlinx.coroutines.coroutineScope
 import kotlin.test.assertEquals
 
-suspend inline fun <I, S> VModel<I, S>.test(intent: I) = coroutineScope { start(intent) }
+suspend inline fun <I, S> ViewModel<I, S>.test(intent: I) = coroutineScope { start(intent) }
 
-suspend fun <I, S> VModel<I, S>.test(intent: I, toHave: S) {
+suspend fun <I, S> ViewModel<I, S>.test(intent: I, toHave: S) {
     test(intent)
-    assertEquals(toHave, ui.value, "Expected ${this::class.simpleName} State to be $toHave but was ${ui.value}")
+    assertEquals(toHave, state.value, "Expected ${this::class.simpleName} State to be $toHave but was ${state.value}")
 }